refactor(taxa): extract create_taxon hierarchy builder into a shared service
Move Command.create_taxon() and get_or_create_root_taxon() out of the
import_taxa management command and into ami.main.services.taxonomy, so
the regional taxa-list service can reuse the same rank-hierarchy builder
instead of re-deriving it. Behaviour is unchanged; import_taxa now calls
the extracted functions instead of defining them locally.
